requirements.txt
Pillow>=3.1.0
pyusb==1.0.0b2